+++ sentry-rails/spec/sentry/rails/active_support_breadcrumbs_spec.rb
@@ -1,13 +1,14 @@
 require "spec_helper"
 
 RSpec.describe "Sentry::Breadcrumbs::ActiveSupportLogger", type: :request do
-  before(:all) do
+  before do
     make_basic_app do |sentry_config|
       sentry_config.breadcrumbs_logger = [:active_support_logger]
+      sentry_config.traces_sample_rate = 1.0
     end
   end
 
-  after(:all) do
+  after do
     require 'sentry/rails/breadcrumb/active_support_logger'
     Sentry::Rails::Breadcrumb::ActiveSupportLogger.detach
     # even though we cleanup breadcrumbs in the rack middleware
@@ -29,13 +30,15 @@ RSpec.describe "Sentry::Breadcrumbs::ActiveSupportLogger", type: :request do
     transport.events = []
   end
 
-  it "captures correct data" do
+  it "captures correct data of exception requests" do
     get "/exception"
 
     expect(response.status).to eq(500)
     breadcrumbs = event.dig("breadcrumbs", "values")
     expect(breadcrumbs.count).to eq(2)
-    expect(breadcrumbs.first["data"]).to include(
+
+    breadcrumb = breadcrumbs.detect { |b| b["category"] == "process_action.action_controller" }
+    expect(breadcrumb["data"]).to include(
       {
         "controller" => "HelloController",
         "action" => "exception",
@@ -44,6 +47,31 @@ RSpec.describe "Sentry::Breadcrumbs::ActiveSupportLogger", type: :request do
         "method" => "GET", "path" => "/exception",
       }
     )
+    expect(breadcrumb["data"].keys).not_to include("headers")
+    expect(breadcrumb["data"].keys).not_to include("request")
+    expect(breadcrumb["data"].keys).not_to include("response")
+  end
+
+  it "captures correct request data of normal requests" do
+    p = Post.create!
+
+    get "/posts/#{p.id}"
+
+    breadcrumbs = event.dig("breadcrumbs", "values")
+
+    breadcrumb = breadcrumbs.detect { |b| b["category"] == "process_action.action_controller" }
+    expect(breadcrumb["data"]).to include(
+      {
+        "controller" => "PostsController",
+        "action" => "show",
+        "params" => { "controller" => "posts", "action" => "show", "id" => p.id.to_s },
+        "format" => "html",
+        "method" => "GET", "path" => "/posts/#{p.id}",
+      }
+    )
+    expect(breadcrumb["data"].keys).not_to include("headers")
+    expect(breadcrumb["data"].keys).not_to include("request")
+    expect(breadcrumb["data"].keys).not_to include("response")
   end
 
   it "ignores exception data" do